add shrinking
